+++ src/hooks.js
@@ -19,7 +19,18 @@ export const HOOKS = {
      */
     createRxSchema: [],
     createRxQuery: [],
-    createRxDocument: []
+    createRxDocument: [],
+    /**
+     * runs before a pouchdb-instance is created
+     * gets pouchParameters as attribute so you can manipulate them
+     * {
+     *   location: string,
+     *   adapter: any,
+     *   settings: object
+     * }
+     * @type {Array}
+     */
+    preCreatePouchDb: []
 };
 
 export function runPluginHooks(hookKey, obj) {