+++ test/unit/plugin.test.js
@@ -5,6 +5,7 @@
 import assert from 'assert';
 import platform from 'detect-browser';
 import RxDB from '../../dist/lib/index';
+import * as util from '../../dist/lib/util';
 import * as humansCollection from '../helper/humans-collection';
 
 describe('plugin.test.js', () => {
@@ -48,6 +49,7 @@ describe('plugin.test.js', () => {
             RxDB.plugin(plugin);
             const col = await humansCollection.create();
             assert.equal(col.database.foo, 'bar_createRxDatabase');
+            col.database.destroy();
         });
         it('createRxCollection', async() => {
             const plugin = {
@@ -61,6 +63,7 @@ describe('plugin.test.js', () => {
             RxDB.plugin(plugin);
             const col = await humansCollection.create();
             assert.equal(col.foo, 'bar_createRxCollection');
+            col.database.destroy();
         });
         it('createRxSchema', async() => {
             const plugin = {
@@ -74,6 +77,7 @@ describe('plugin.test.js', () => {
             RxDB.plugin(plugin);
             const col = await humansCollection.create();
             assert.equal(col.schema.foo, 'bar_createRxSchema');
+            col.database.destroy();
         });
         it('createRxQuery', async() => {
             const plugin = {
@@ -88,6 +92,7 @@ describe('plugin.test.js', () => {
             const col = await humansCollection.create();
             const query = col.find();
             assert.equal(query.foo, 'bar_createRxQuery');
+            col.database.destroy();
         });
         it('createRxDocument', async() => {
             const plugin = {
@@ -102,6 +107,29 @@ describe('plugin.test.js', () => {
             const col = await humansCollection.create(5);
             const doc = await col.findOne().exec();
             assert.equal(doc.foo, 'bar_createRxDocument');
+            col.database.destroy();
+        });
+        it('preCreatePouchDb', async() => {
+            const collectionName = util.randomCouchString(10);
+            const plugin = {
+                rxdb: true,
+                hooks: {
+                    preCreatePouchDb: pouchDbParameters => {
+                        if (pouchDbParameters.location.includes(collectionName)) {
+                            // only do sth at this specific collection-pouch
+                            pouchDbParameters.location = pouchDbParameters.location + 'foobar';
+                        }
+                    }
+                }
+            };
+            RxDB.plugin(plugin);
+            const col = await humansCollection.create(0, collectionName);
+            const pouchInstance = col.pouch;
+            assert.ok(pouchInstance);
+
+            const info = await pouchInstance.info();
+            assert.ok(info.db_name.includes('foobar'));
+            col.database.destroy();
         });
     });
 });